[Exchnages] replace Huobi by HTX
