+++ Trading/Exchange/htx/metadata.json
@@ -1,0 +1,6 @@
+{
+  "version": "1.2.0",
+  "origin_package": "OctoBot-Default-Tentacles",
+  "tentacles": ["Htx"],
+  "tentacles-requirements": []
+}
