+++ Trading/Exchange/htx/resources/htx.md
@@ -1,0 +1,1 @@
+HTX is a basic RestExchange adaptation for HTX exchange.
