+++ Trading/Exchange/htx_websocket_feed/metadata.json
@@ -1,0 +1,6 @@
+{
+  "version": "1.2.0",
+  "origin_package": "OctoBot-Default-Tentacles",
+  "tentacles": ["HtxCCXTWebsocketConnector"],
+  "tentacles-requirements": []
+}
